+++ NEWS.md
@@ -29,12 +29,15 @@
 * `split_by()`: Now also outputs NA values. (29.03.2026)
 * `import_data()`, `import_multi()`, `export_data()`, `export_multi()`: Now can import and export csv files with file extension txt. (30.03.2026)
 * `round_values()`: Now is also able to round in multiples of a provided value. (09.04.2026)
+* `combine_into_worbook()`: Now can handle `export_with_style()` results. (10.04.2026)
+* `excel_output_style()`: Now accepts multiple title and footnote colors, sizes and boldings. (10.04.2026)
 
 ### Changed functionality
 
 * `summarise_plus()`: Instead of aborting when no values are passed, the function now generates a variable to output unweighted results. (26.03.2026)
 * `Unit tests`: Moved unit tests completely from `testthat` to `tinytest`. `testthat` repeatedly got in the way with its own environment, `tinytest` offers more natural conditions and works faster. (31.01.2026)
 * `transpose_plus()`: Removed macro variable usage. (09.04.2026)
+* `export_with_style()`: Now returns a list with table, workbook and meta information, which can be used with `combine_into_worbook()`. (10.04.2026)
 
 ### Fixed
 
@@ -51,6 +54,8 @@
 * `split_by()`: Should now work inside custom functions. (29.03.2026)
 * `retain_value()`, `retain_sum()`: On error now return NA instead of the whole data frame. (31.01.2026)
 * `transpose_plus()`: Wide to long transposition now working again as intended. (09.04.2026)
+* `any_table()`: Without column variables each variable in the column header received a "_ " at the end, which is not the case anymore. (10.04.2026)
+* `sort_plus()`: When formats where used for sorting, the temporary sorting variables where always put up front instead of inserted in sequential order. This is fixed now. (10.04.2026)
 
 ### Optimization
 
@@ -70,6 +75,8 @@
 * `any_table()`, `frequencies()`, `crosstabs()`, `export_with_style()`: Added missing global options function in the `see also` section. (25.03.2026)
 * `import_multi()`, `export_multi()`: Added some messages to show progress. (28.03.2026)
 * Added repository to DeepWiki and added badge to README. (29.03.2026)
+* `any_table()`, `export_with_style()`: Added examples for `combine_into_worbook()`. (10.04.2026)
+* `any_table()`: When row header labels are suppressed, the excess slashes are now removed. (10.04.2026)
 
 # qol 1.2.2